Se implementa control por sessiones, validaciones para inicio de sesion

+++ ProyectoFinal/controllers/productoController.js
@@ -2,9 +2,18 @@ require('dotenv').config();
 const {Cuentas, Productos} = require('../models/productoDB');
 
 const agregarProductoGET = (req, res) => {
-    res.render('agregar-producto', {
-        titulo: "Agregar producto"
-    });
+
+    try {
+        if(req.session.logueado === true){
+            res.render('agregar-producto', {
+                titulo: "Agregar producto"
+            });
+        } else {
+            res.redirect('/login');
+        }
+    } catch (error) {
+        console.log('[ERROR]' + error);
+    };
 };
 
 const agregarProductoPOST = async (req, res) => {
@@ -26,7 +35,7 @@ const agregarProductoPOST = async (req, res) => {
                 titulo: "Agregar producto"
             });
         } else {
-            res.redirect('/login')
+            res.redirect('/login');
         };
     } catch (error) {
         console.log('[ERROR]' + error);
@@ -37,25 +46,29 @@ const editarProductoGET = async (req, res) => {
     const idProducto = req.params.id
 
     try {
-        const productoFind = await Productos.findOne({ 
-            where: {
-                id: idProducto
-            } 
-        })
-
-        if( !productoFind ) {
-            res.send(`
-                <h1>No existe el producto con id ${idProducto}</h1>
-                <a href="/admin"> Ver listado de productos</a>
-            `)
+        if(req.session.logueado === true){
+            const productoFind = await Productos.findOne({ 
+                where: {
+                    id: idProducto
+                } 
+            })
+    
+            if( !productoFind ) {
+                res.send(`
+                    <h1>No existe el producto con id ${idProducto}</h1>
+                    <a href="/admin"> Ver listado de productos</a>
+                `)
+            } else {
+                res.render('editar-producto', {
+                    titulo: "Vista del administrador",
+                    productos: productoFind
+                });    
+            };
         } else {
-            res.render('editar-producto', {
-                titulo: "Vista del administrador",
-                productos: productoFind
-            });    
-        };
+            res.redirect('/login');
+        }
     } catch (error) {
-        console.log(error)
+        console.log(error);
     };
 };
 
@@ -64,21 +77,25 @@ const editarProductoPOST = async (req, res) => {
     const producto = req.body;
     
     try {
-        const nuevoProducto = await Productos.update({
-            nombre: producto.nombre,
-            descripcion: producto.descripcion,
-            caracteristica: producto.caracteristica,
-            precio: producto.precio,
-            stock: producto.stock,
-            rutaImagen: producto.rutaImagen,
-            destacado: producto.destacado
-        }, 
-        {
-            where: { id: idProducto }
-        })
-        console.log('SE MODIFICO CORRECTAMENTE')
-        // console.log(nuevoProducto);
-        res.redirect('/admin');
+        if( req.session.logueado === true ){
+            const nuevoProducto = await Productos.update({
+                nombre: producto.nombre,
+                descripcion: producto.descripcion,
+                caracteristica: producto.caracteristica,
+                precio: producto.precio,
+                stock: producto.stock,
+                rutaImagen: producto.rutaImagen,
+                destacado: producto.destacado
+            }, 
+            {
+                where: { id: idProducto }
+            })
+            console.log('SE MODIFICO CORRECTAMENTE')
+            // console.log(nuevoProducto);
+            res.redirect('/admin');
+        } else {
+            res.redirect('/login');
+        };
     } catch (error) {
         console.log('[ERROR]' + error);
     };
@@ -87,14 +104,22 @@ const editarProductoPOST = async (req, res) => {
 const borrarProductoGET = async (req, res) => {
     const idProducto = req.params.id
 
-    const productoDelete = await Productos.destroy({
-        where: {
-            id: idProducto
+    try {
+        if( req.session.logueado === true ) {
+            const productoDelete = await Productos.destroy({
+                where: {
+                    id: idProducto
+                }
+            });
+            console.log('PRODUCTO ELIMINADO');
+            // res.send('Producto eliminado')
+            res.redirect('/admin');
+        } else {
+            res.redirect('/login');
         }
-    });
-    console.log('PRODUCTO ELIMINADO');
-    // res.send('Producto eliminado')
-    res.redirect('/admin');
+    } catch (error) {
+        console.log('[ERROR]' + error);
+    };    
 };
